添加压测程序

+++ src/infra/health/HealthService.cpp
@@ -142,7 +142,7 @@ void HealthService::fillMaintenanceStats(HealthSnapshot& snapshot){
     //判断长期没有成功
     const int64_t nowMs=currentEpochMs();
     if(maintenanceSnapshot.lastSuccessAtMs>0&&maintenanceIntervalMs_>0){
-        const uint64_t maintenanceThresholdMs =3ULL * static_cast<uint64_t>(maintenanceIntervalMs_);
+        const uint64_t maintenanceThresholdMs =3LL *maintenanceIntervalMs_;
         if(nowMs>=maintenanceSnapshot.lastSuccessAtMs){
             const auto elapsedMs = static_cast<uint64_t>(nowMs - maintenanceSnapshot.lastSuccessAtMs);
             snapshot.maintenanceStale =elapsedMs > maintenanceThresholdMs;
@@ -150,7 +150,7 @@ void HealthService::fillMaintenanceStats(HealthSnapshot& snapshot){
     }
     //判断运行时间过长
     if(maintenanceSnapshot.running){
-        const uint64_t maintenanceThresholdMs =2ULL * static_cast<uint64_t>(maintenanceIntervalMs_);
+        const int64_t maintenanceThresholdMs =2ULL * static_cast<uint64_t>(maintenanceIntervalMs_);
         if((currentEpochMs()-maintenanceSnapshot.lastRunAtMs)>maintenanceThresholdMs){
             snapshot.maintenanceRunningTooLong=true;
         }
